From the Love-O-Meter Project
I added a button to it to "control" programmatically the functioning of the board. When the button is "off" the program ignores readings from the temperature sensor.

Without altering the board with an external power switch there is no way to give the board a true power switch using existing components in the starter kit.

Serial.print(""); lines added to aid visibility for testing.